fix: 認証画面 デザイン修正
