+++ frontend/src/App.css
@@ -182,3 +182,20 @@
     border-right-color: var(--border);
   }
 }
+
+.page {
+  /* min-height: 100vh; */
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+  padding-top: 100px;
+  padding-bottom: 40px;
+}
+
+.container {
+  background: #1e1e1e;
+  border: 1px solid #444;
+  border-radius: 12px;
+  padding: 24px;
+  box-sizing: border-box;
+}
