+++ frontend/src/components/FormField.css
@@ -1,0 +1,24 @@
+.form-group {
+  display: flex;
+  flex-direction: column;
+}
+
+.form-label {
+  margin-bottom: 6px;
+  font-size: 14px;
+}
+
+.form-input {
+  /* width: 100%; */
+  padding: 10px;
+  background: #2a2a2a;
+  border: 1px solid #444;
+  border-radius: 6px;
+  color: #fff;
+  font-size: 18px;
+}
+
+input:focus {
+  outline: none;
+  border: 1px solid #b1b1b1;
+}
